reafactor(tauri.js): move command helpers to api dir
