+++ cli/tauri.js/api/build.js
@@ -1,9 +1,9 @@
 module.exports = config => {
-  const { tauriDir } = require('./app-paths')
+  const { tauriDir } = require('../helpers/app-paths')
   const merge = require('webpack-merge')
   const Runner = require('../runner')
   const tauri = new Runner({ modeDir: tauriDir })
-  const tauriConfig = require('./tauri-config')(
+  const tauriConfig = require('../helpers/tauri-config')(
     merge(
       {
         ctx: {
