+++ cli/tauri.js/api/dev.js
@@ -1,9 +1,9 @@
 module.exports = config => {
-  const { tauriDir } = require('./app-paths')
+  const { tauriDir } = require('../helpers/app-paths')
   const Runner = require('../runner')
   const merge = require('webpack-merge')
   const tauri = new Runner()
-  const tauriConfig = require('./tauri-config')(
+  const tauriConfig = require('../helpers/tauri-config')(
     merge(
       {
         ctx: {
